Make Configuration extensible
