+++ tests/Optimizer/DefaultConfigurationTest.php
@@ -9,21 +9,21 @@ use AmpProject\Tests\TestCase;
 /**
  * Test the configuration storage and validation.
  *
- * @covers \AmpProject\Optimizer\Configuration
+ * @covers \AmpProject\Optimizer\DefaultConfiguration
  * @package ampproject/amp-toolbox
  */
-final class ConfigurationTest extends TestCase
+final class DefaultConfigurationTest extends TestCase
 {
 
     /**
      * Test whether we can retrieve the default configuration.
      *
-     * @covers \AmpProject\Optimizer\Configuration::has()
-     * @covers \AmpProject\Optimizer\Configuration::get()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::has()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::get()
      */
     public function testDefaultConfiguration()
     {
-        $configuration = new Configuration();
+        $configuration = new DefaultConfiguration();
         $this->assertTrue($configuration->has(Configuration::KEY_TRANSFORMERS));
         $this->assertFalse($configuration->has('unknown_key'));
         $this->assertEquals(Configuration::DEFAULT_TRANSFORMERS, $configuration->get(Configuration::KEY_TRANSFORMERS));
@@ -32,12 +32,12 @@ final class ConfigurationTest extends TestCase
     /**
      * Test whether we can add to the default configuration.
      *
-     * @covers \AmpProject\Optimizer\Configuration::has()
-     * @covers \AmpProject\Optimizer\Configuration::get()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::has()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::get()
      */
     public function testUserProvidedConfigurationCanAddKeys()
     {
-        $configuration = new Configuration(['custom_key' => 'custom_value']);
+        $configuration = new DefaultConfiguration(['custom_key' => 'custom_value']);
         $this->assertTrue($configuration->has(Configuration::KEY_TRANSFORMERS));
         $this->assertTrue($configuration->has('custom_key'));
         $this->assertEquals(Configuration::DEFAULT_TRANSFORMERS, $configuration->get(Configuration::KEY_TRANSFORMERS));
@@ -47,12 +47,12 @@ final class ConfigurationTest extends TestCase
     /**
      * Test whether we can override the default configuration.
      *
-     * @covers \AmpProject\Optimizer\Configuration::has()
-     * @covers \AmpProject\Optimizer\Configuration::get()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::has()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::get()
      */
     public function testUserProvidedConfigurationCanOverrideKeys()
     {
-        $configuration = new Configuration([Configuration::KEY_TRANSFORMERS => ['my_transformer']]);
+        $configuration = new DefaultConfiguration([Configuration::KEY_TRANSFORMERS => ['my_transformer']]);
         $this->assertTrue($configuration->has(Configuration::KEY_TRANSFORMERS));
         $this->assertEquals(['my_transformer'], $configuration->get(Configuration::KEY_TRANSFORMERS));
     }
@@ -60,11 +60,11 @@ final class ConfigurationTest extends TestCase
     /**
      * Test whether unknown keys throw an exception.
      *
-     * @covers \AmpProject\Optimizer\Configuration::get()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::get()
      */
     public function testUnknownKeyThrowsException()
     {
-        $configuration = new Configuration();
+        $configuration = new DefaultConfiguration();
         $this->expectException(UnknownConfigurationKey::class);
         $configuration->get('unknown_key');
     }
@@ -72,24 +72,24 @@ final class ConfigurationTest extends TestCase
     /**
      * Test whether invalid keys throw an exception.
      *
-     * @covers \AmpProject\Optimizer\Configuration::validateConfigurationKeys()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::validateConfigurationKeys()
      */
     public function testInvalidTransformersTypeThrowsException()
     {
         $this->expectException(InvalidConfigurationValue::class);
         $this->expectExceptionMessage("The configuration key 'transformers' expected a value of type 'array', got 'integer' instead.");
-        new Configuration([Configuration::KEY_TRANSFORMERS => 42]);
+        new DefaultConfiguration([Configuration::KEY_TRANSFORMERS => 42]);
     }
 
     /**
      * Test whether invalid sub-keys throw an exception.
      *
-     * @covers \AmpProject\Optimizer\Configuration::validateConfigurationKeys()
+     * @covers \AmpProject\Optimizer\DefaultConfiguration::validateConfigurationKeys()
      */
     public function testInvalidTransformersSubTypeThrowsException()
     {
         $this->expectException(InvalidConfigurationValue::class);
         $this->expectExceptionMessage("The configuration value '2' for the key 'transformers' expected a value of type 'string', got 'integer' instead.");
-        new Configuration([Configuration::KEY_TRANSFORMERS => ['first_one_is_good', 'second_one_is_good_as_well_but...', 42]]);
+        new DefaultConfiguration([Configuration::KEY_TRANSFORMERS => ['first_one_is_good', 'second_one_is_good_as_well_but...', 42]]);
     }
 }
